requirements.txt
# Core
python-dotenv>=1.0.0
loguru>=0.7.2

# HTTP client (bridge communication)
httpx>=0.27.0

# WebSocket (event stream + host WS server)
websockets>=13.0

# AI providers
anthropic>=0.40.0
openai>=1.50.0

# Host API server
fastapi>=0.115.0
uvicorn>=0.32.0
pydantic>=2.9.0

# Deploy script
paramiko>=3.5.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
respx>=0.22.0

# Development
black>=24.0.0
flake8>=7.0.0
mypy>=1.13.0
